Initial commit: Add existing project files
This commit adds a reduced and simplified version of the OSEAX forecasting project from my master's thesis. It includes code to run the LSTM model with standard hyperparameters, capable of forecasting over 1, 5, 22, or 261 business days. The code supports forecasts on either a univariate dataset (OSEAX history only) or a multivariate dataset with 10 features.